Create services
- Create `MessageTransferService`
- Write JavaDoc for test of `TelegramMessageTransferService`
- Create test for `TelegramMessageQueryService`
- Realize MessageQueryService

+++ src/main/java/com/instrumentisto/timebot/service/MessageTransferService.java
@@ -22,7 +22,7 @@ public interface MessageTransferService {
      * @param id value of ID message which need to get.
      * @return {@link Message} value.
      */
-    Message findById(long id);
+    Message findById(int id);
 
     /**
      * Returns list of all {@link Message} objects from repository.
